fix segment timing info (#498)
* fix segment timing info

* fix duration

* last segment is allowed to be short

* fix duration

* fix live + manifest

+++ pkg/pipeline/watch.go
@@ -290,7 +290,7 @@ func (c *Controller) handleMessageElement(msg *gst.Message) error {
 	return nil
 }
 
-func getSegmentParamsFromGstStructure(s *gst.Structure) (filepath string, time int64, err error) {
+func getSegmentParamsFromGstStructure(s *gst.Structure) (filepath string, time uint64, err error) {
 	loc, err := s.GetValue(fragmentLocation)
 	if err != nil {
 		return "", 0, err
@@ -309,7 +309,7 @@ func getSegmentParamsFromGstStructure(s *gst.Structure) (filepath string, time i
 		return "", 0, errors.ErrGstPipelineError(errors.New("invalid type for time"))
 	}
 
-	return filepath, int64(ti), nil
+	return filepath, ti, nil
 }
 
 func getFirstSampleMetadataFromGstStructure(s *gst.Structure) (startDate time.Time, err error) {
